+++ hexen2/snd_mem.c
@@ -225,7 +225,7 @@ static void DumpChunks(void)
 GetWavinfo
 ============
 */
-wavinfo_t GetWavinfo (char *name, byte *wav, int wavlength)
+wavinfo_t GetWavinfo (char *name, byte *wav, size_t wavlength)
 {
 	wavinfo_t	info;
 	int	i;
