+++ hexen2/sound.h
@@ -137,7 +137,7 @@ extern int		snd_blocked;
 void S_LocalSound (char *sound);
 sfxcache_t *S_LoadSound (sfx_t *s);
 
-wavinfo_t GetWavinfo (char *name, byte *wav, int wavlength);
+wavinfo_t GetWavinfo (char *name, byte *wav, size_t wavlength);
 
 void SND_InitScaletable (void);
